backend/requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
plotly
fpdf
streamlit-authenticator
flask
flask-cors
google-generativeai
python-dotenv



